Update Scurl library for handling multiple CURL requests
This commit includes changes in the Scurl library to allow it to handle multiple CURL requests. Changes include modifying the 'parameters' property in 'Scurl_Request' from string type to an array type and updating methods in 'Curl' trait accordingly to work with the changes. Also, a new 'getRange' utility function is added to 'Utils' trait for getting a specific range based on an id and a divider.